Create TagNewsList screen.

+++ app/src/main/java/my/dzeko/footapp/database/dao/NewsDao.kt
@@ -29,5 +29,5 @@ interface NewsDao {
     fun getLatestNewsDate(): Long?
 
     @Query("Select title, T1.id, summary, date From News As T1 Inner Join NewsTag as T2 On T1.id=T2.newsId Where T2.tagId=:tagId Order By date Desc")
-    fun getNewsByTagId(tagId: Int): DataSource.Factory<Int, NewsSummary>
+    fun getNewsByTagId(tagId: Long): DataSource.Factory<Int, NewsSummary>
 }
