+++ app/src/main/java/my/dzeko/footapp/repository/news/NewsRepository.kt
@@ -34,7 +34,7 @@ class NewsRepository @Inject constructor(
         mLocalRepo.save(news)
     }
 
-    suspend fun getNewsByTagId(tagId: Int): LiveData<PagedList<NewsSummary>> {
+    suspend fun getNewsByTagId(tagId: Long): LiveData<PagedList<NewsSummary>> {
         return withContext(Dispatchers.IO) {
             mLocalRepo.getNewsByTagId(tagId)
         }
